big bindings refactor, leading whitespace
