+++ assembly/error/index.ts
@@ -1,41 +1,5 @@
-
-/**
- * Obtain the length of an error string.
- *
- * @param {u64} id - The id of the error.
- * @returns {usize} The length of the string.
- */
-
+import { error } from "../bindings";
 import { add_finalize, LunaticManaged } from "../util";
-
-// @ts-ignore: external is valid here
-@external("lunatic::error", "string_size")
-export declare function string_size(id: u64): usize;
-
-
-/**
- * Write the utf8 string into memory.
- *
- * @param {u64} id - The error id.
- * @param {usize} ptr [*mut u8] The pointer to memory where it will be written.
- */
-// @ts-ignore
-@external("lunatic::error", "to_string")
-export declare function to_string(id: u64, ptr: usize): void;
-
-/**
- * Drop the error
- *
- * @param {u64} id - The error id.
- */
-// @ts-ignore
-@external("lunatic::error", "drop")
-export declare function drop_error(id: u64): void;
-
-export const enum err_code {
-  Success,
-  Fail,
-}
 
 export let err_str: string | null = "";
 
@@ -47,10 +11,10 @@ export let err_str: string | null = "";
  * @returns The error string.
  */
 export function getError(id: u64): string {
-  let len = string_size(id);
+  let len = error.string_size(id);
   let ptr = heap.alloc(len);
-  to_string(id, ptr);
-  drop_error(id);
+  error.to_string(id, ptr);
+  error.drop_error(id);
   let value = String.UTF8.decodeUnsafe(ptr, len, false);
   heap.free(ptr);
   return value;
@@ -74,7 +38,7 @@ export class Result<T> extends LunaticManaged {
     let errId = this.errId;
     if (errId === u64.MAX_VALUE) return "";
 
-    let errStr = this.errStr; 
+    let errStr = this.errStr;
 
     if (errStr == null) {
       this.dropped = true;
@@ -88,11 +52,11 @@ export class Result<T> extends LunaticManaged {
   dispose(): void {
     this.drop();
   }
-  
+
   /** Dispose the error string if it was allocated. */
   drop(): void {
     if (!this.dropped && this.errId != u64.MAX_VALUE) {
-      drop_error(this.errId);
+      error.drop_error(this.errId);
       this.dropped = true;
     }
   }
