+++ assembly/util/index.ts
@@ -27,3 +27,22 @@ export abstract class LunaticManaged {
 
   abstract dispose(): void;
 }
+
+export const enum ReceiveType {
+  DataMessage = 0,
+  SignalMessage = 1,
+  Timeout = 9027,
+}
+
+export const enum MessageType {
+  None = 0,
+  Signal = 1,
+  Error = 2,
+  Value = 3,
+}
+
+/** Success enum to describe the results of syscalls. The value `0` is successful. */
+export const enum err_code {
+  Success,
+  Fail,
+}
